docs(ai): add outputs from arch session for project recovery
These are living docs and context for AI and humans alike, focusing
on getting this to a steady state with all the incorrect choices
sorted out and a new agile AI agent team based workflow integrated.

- Add artifacts generated by the arch agent after a day of back and forth.
  This adds to the existing outputs from BA and PM agents.
- prd-resolver.md was updated by the PM agent when we found discrepancies
- Epics 1, 2 refined and new stories added
- Lots of planning and iteration on current blockers, and a full code base review
- Service layer entierily refactored along with schemas (no code changes, just docs and artifacts)
- Rest of arch artifacts, focusing on stabilizing the project.

There's much to be improved, but this should get us to a state
where now hopefully have enough context and structure to drive
forward with the SM and Dev agents.

Next stage is SM agent <-> Dev agent <-> human collaboration in IDE
to drill dwon to stories one at a time.
See docs/templates/workflow-diagram.md